Move design-related functions into separate file
