+++ monitor/src/signal_trace.rs
@@ -2,7 +2,7 @@
 // released under MIT License
 // author: Kevin Laeufer <laeufer@cornell.edu>
 
-use crate::{Design, Instance};
+use crate::{Instance, designs::Design};
 use baa::BitVecValue;
 use protocols::ir::SymbolId;
 use rustc_hash::FxHashMap;
